Backend/FlaskBackend/requirements.txt
flask[async]
flask_cors
openai
pycryptodome
pyTelegramBotAPI
aptos_sdk
pillow
asyncio
gunicorn
python-dotenv
telethon
pymongo